+++ src/Shipment/Changelog.md
@@ -1,5 +1,10 @@
 # Vanilo Shipment Module Changelog
 
+## Unreleased
+##### 2022-03-XX
+
+- Added `actives()` and `inactives()` scopes to the Carrier model
+
 ## 3.0.0
 ##### 2022-02-28